+++ python/bullmq/queue.py
@@ -1,8 +1,14 @@
 import redis.asyncio as redis
+from typing import TypedDict
 
 from bullmq.scripts import Scripts
 from bullmq.job import Job
 from bullmq.redis_connection import RedisConnection
+
+class RetryJobsOpts(TypedDict):
+    state : str
+    count : int
+    timestamp: int
 
 class Queue:
     """
@@ -43,7 +49,7 @@ class Queue:
         return self.scripts.pause(False)
 
     async def isPaused(self):
-        pausedKeyExists = await self.conn.hexists(self.opts.get("prefix") or "bull" + ":" + self.name + ":meta", "paused")
+        pausedKeyExists = await self.client.hexists(self.opts.get("prefix") or "bull" + ":" + self.name + ":meta", "paused")
         return pausedKeyExists == 1
 
     """
@@ -56,6 +62,25 @@ class Queue:
             cursor = await self.scripts.obliterate(1000, force)
             if cursor == 0 or cursor == None or cursor == "0":
                 break
+
+    """
+        Retry all the failed jobs.
+    """
+    async def retryJobs(self, opts: RetryJobsOpts = {}):
+        while True:
+            cursor = await self.scripts.retryJobs(
+                opts.get("state"),
+                opts.get("count"),
+                opts.get("timestamp")
+            )
+            if cursor == 0 or cursor == None or cursor == "0":
+                break
+
+    """
+        Trim the event stream to an approximately maxLength.
+    """
+    def trimEvents(self, maxLength: int):
+        return self.client.xtrim(self.opts.get("prefix") or "bull" + ":" + self.name + ":events", "MAXLEN", "~", maxLength)
 
     """
         Closes the queue and the underlying connection to Redis.